no more HTTP 500 when external search service calls fail
